+++ src/providers/twitch/TwitchIrcServer.hpp
@@ -15,8 +15,6 @@ namespace chatterino {
 class Settings;
 class Paths;
 class TwitchChannel;
-class BttvLiveUpdates;
-class SeventvEventAPI;
 class BttvEmotes;
 class FfzEmotes;
 class SeventvEmotes;
@@ -34,9 +32,6 @@ public:
 
     virtual void dropSeventvChannel(const QString &userID,
                                     const QString &emoteSetID) = 0;
-
-    virtual std::unique_ptr<BttvLiveUpdates> &getBTTVLiveUpdates() = 0;
-    virtual std::unique_ptr<SeventvEventAPI> &getSeventvEventAPI() = 0;
 
     virtual const IndirectChannel &getWatchingChannel() const = 0;
     virtual void setWatchingChannel(ChannelPtr newWatchingChannel) = 0;
@@ -102,13 +97,7 @@ private:
     const ChannelPtr automodChannel;
     IndirectChannel watchingChannel;
 
-    std::unique_ptr<BttvLiveUpdates> bttvLiveUpdates;
-    std::unique_ptr<SeventvEventAPI> seventvEventAPI;
-
 public:
-    std::unique_ptr<BttvLiveUpdates> &getBTTVLiveUpdates() override;
-    std::unique_ptr<SeventvEventAPI> &getSeventvEventAPI() override;
-
     const IndirectChannel &getWatchingChannel() const override;
     void setWatchingChannel(ChannelPtr newWatchingChannel) override;
     ChannelPtr getWhispersChannel() const override;